Move switch-tests to their own package
